add a built-in global runtime providing a basic pair data structure
 * OBJECT which is class object {}
 * pair(a, b) which uses OBJECT() and .first/.second to form a pair
 * pair_first(p) and pair_second(p) for read access to the pair halves

These are implemented in the VM opcodes added to the global namespace
during the vm init process.

The underlying philosophy here is not to provide compiler support for
class and property access, but something that's even simpler to implement

The full code from chapter 27 of Crafting Interpreters or even 28
could have easilly been ported over, but the idea here is to enforce
a more rigid subset of Lox that a bootstrapped implementation can more
easilly implement.

This implementation can thus act as an enforced reference implementation.

By enforcing this, alternative implementations could also implement
pair(), pair_first(), and pair_second() in more native/low-level manners